refactor(debug): rename apps/client to apps/debug
It's a debugger for rssCloud/WebSub servers, not an SDK-style client, so
the old name was misleading. Renames the directory, package
(@rsscloud/client-app -> @rsscloud/debug), entry file (client.js ->
debug.js), env var (CLIENT_FETCH_ALLOW_CIDRS -> DEBUG_FETCH_ALLOW_CIDRS),
page branding, root npm script, Docker image tag, and release-please
tracking. Leaves untouched anything naming the actual rssCloud/WebSub
protocol client (lib/client.js, createRssCloudClient/createWebSubClient,
and CONTEXT.md's "Client" domain role), since that's a different concept.
