+++ apps/debug/.env.example
@@ -1,10 +1,10 @@
-# rssCloud test harness — environment template.
+# rssCloud debug harness — environment template.
 #
-# Copy to apps/client/.env (gitignored) and adjust:
-#     cp apps/client/.env.example apps/client/.env
+# Copy to apps/debug/.env (gitignored) and adjust:
+#     cp apps/debug/.env.example apps/debug/.env
 #
-# The harness reads .env from its own directory (apps/client) via dotenv, so
-# this file must live here, not at the repo root. `pnpm client` runs from here.
+# The harness reads .env from its own directory (apps/debug) via dotenv, so
+# this file must live here, not at the repo root. `pnpm debug` runs from here.
 #
 # The values below are tuned for LOCAL DEVELOPMENT alongside the rssCloud hub
 # (apps/server), which listens on http://localhost:5337 by default.
@@ -29,7 +29,7 @@ HUB_SERVER_URL=http://localhost:5337
 # >>> PRODUCTION: delete this line before deploying publicly. This harness is
 # >>> meant to hit arbitrary third-party feeds/hubs when deployed live —
 # >>> loopback exemptions have no place there.
-CLIENT_FETCH_ALLOW_CIDRS=127.0.0.0/8,::1/128
+DEBUG_FETCH_ALLOW_CIDRS=127.0.0.0/8,::1/128
 
 # --- Optional overrides (defaults shown; uncomment to change) --------------
 # REQUEST_TIMEOUT=4000                 # outbound fetch timeout (ms), SSRF-guarded
