+++ apps/debug/Dockerfile
@@ -5,7 +5,7 @@ RUN corepack enable
 WORKDIR /app
 
 COPY package.json pnpm-workspace.yaml pnpm-lock.yaml ./
-COPY apps/client/package.json apps/client/
+COPY apps/debug/package.json apps/debug/
 COPY packages/xml-rpc/package.json packages/xml-rpc/
 COPY packages/core/package.json packages/core/
 
@@ -19,15 +19,15 @@ RUN pnpm --filter "@rsscloud/core..." run build
 
 FROM base AS dependencies
 
-RUN pnpm install --frozen-lockfile --filter "@rsscloud/client-app..." --prod --ignore-scripts
+RUN pnpm install --frozen-lockfile --filter "@rsscloud/debug..." --prod --ignore-scripts
 
 FROM dependencies AS runtime
 
-COPY --chown=node:node apps/client apps/client
+COPY --chown=node:node apps/debug apps/debug
 COPY --from=build --chown=node:node /app/packages/xml-rpc/dist packages/xml-rpc/dist
 COPY --from=build --chown=node:node /app/packages/core/dist packages/core/dist
 
-WORKDIR /app/apps/client
+WORKDIR /app/apps/debug
 
 # Default listen port (config.js's PORT default). Documentation only —
 # override PORT at runtime and publish the matching port if you change it.
