+++ apps/debug/lib/client.js
