+++ apps/debug/lib/discover.js
