+++ apps/debug/lib/guarded-fetch.js
