+++ apps/debug/public/app.js
