+++ apps/server/.env.example
@@ -6,19 +6,19 @@
 # The server reads .env from its own directory (apps/server) via dotenv, so this
 # file must live here, not at the repo root. `pnpm start` runs from here.
 #
-# The values below are tuned for LOCAL DEVELOPMENT alongside the client harness
-# (apps/client), which serves its feed and WebSub callback on http://localhost:9000.
+# The values below are tuned for LOCAL DEVELOPMENT alongside the debug harness
+# (apps/debug), which serves its feed and WebSub callback on http://localhost:9000.
 
 # --- Hub identity ---------------------------------------------------------
-# The client harness hardcodes the hub at http://localhost:5337, so keep these
+# The debug harness hardcodes the hub at http://localhost:5337, so keep these
 # matching for local dev. HUB_URL defaults to http://${DOMAIN}:${PORT}${WEBSUB_PATH}.
 DOMAIN=localhost
 PORT=5337
 
 # --- SSRF egress guard: allow loopback for local dev ----------------------
 # The egress guard is ALWAYS ON and refuses loopback/private targets by default.
-# For local dev the hub must reach the client on loopback:
-#   FETCH allowlist    -> topic re-fetch on ping (the client's feed)
+# For local dev the hub must reach the debug harness on loopback:
+#   FETCH allowlist    -> topic re-fetch on ping (the debug harness's feed)
 #   CALLBACK allowlist -> WebSub verification GET + delivery, and the rssCloud
 #                         challenge GET / notify sent to the callback
 # Covers 127.x (IPv4) and ::1 (IPv6) since `localhost` may resolve to either.
